Rename properties, use static const global variables instead of #define
The properties menuSide and options are renamed to menuLocation and menuOptions, respectively, to better communicate their usage.

This commit moves the global preprocessor define kSidebarWidth to the implementation file and refactors it to be a static const type. This allows a much better debugging experience, because the compiler can check the types wherever the constants are used.