Do not invoke orientSideMenuFromStatusBar when configuring MFSideMenuManager
Invoking orientSideMenuFromStatusBar in -[MFSideMenuManager configureWithNavigationController:sideMenuController:menuSide:options:] causes unwanted side effects when the side menu controller is a UINavigationController. In this situation, the frames of the top view controller's view and navigation bar are not set correctly. This was introduced in commit 7afb3dccaf39969a6f5956814ef2a95f5d218b9b. Actually, I am not sure what the purpose of this invocation is, since everything works as expected in the sample project even when the invocation is missing.